backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bodyspec-learn"
version = "1.0.0"
description = "DEXA Education Platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.9",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    "weasyprint>=62.0",
    "jinja2>=3.1.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0.0",
    "markdown>=3.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "flake8>=7.0.0",
    "flake8-quotes>=3.4.0",
    "flake8-isort>=6.1.0",
    "mypy>=1.11.0",
    "isort>=5.13.0",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.isort]
profile = "black"
line_length = 100
skip = [".venv", "venv", "__pycache__"]

[tool.mypy]
python_version = "3.12"
disallow_untyped_defs = true
disallow_incomplete_defs = true
ignore_missing_imports = true
